use sklearn for confusion matrix, recall+precision added

+++ util/cmd_parser.py
@@ -49,9 +49,9 @@ def add_common_arguments(parser):
 
     parser.add_argument('--trials_slices', type=int, default=eeg_config.TRIAL_SLICES,
                         help=f'Will slice every Trial into n x Trials (default: {eeg_config.TRIAL_SLICES})')
-    parser.add_argument('--tmin', type=int, default=eeg_config.TMIN,
+    parser.add_argument('--tmin', type=float, default=eeg_config.TMIN,
                         help=f'Start Time of every Trial Epoch (default: {eeg_config.TMIN})')
-    parser.add_argument('--tmax', type=int, default=eeg_config.TMAX,
+    parser.add_argument('--tmax', type=float, default=eeg_config.TMAX,
                         help=f'End Time of every Trial Epoch (default: {eeg_config.TMAX})')
